# openapi.yaml
openapi: 3.0.3
info:
  title: Fortpesa Payment Platform API
  version: 1.0.0
  description: >
    Merchant-facing payment backend integrating Fortpesa (M-Pesa STK push)
    for payment collection, reconciliation and ledgering. See
    docs/fortpesa.md for which provider behaviors are confirmed versus
    assumed pending Fortpesa's own developer documentation.
servers:
  - url: https://api.your-domain.example
    description: Production
  - url: http://localhost:3000
    description: Local development
security:
  - bearerAuth: []
paths:
...
  /api/v1/payments/{id}/refunds:
    post:
      summary: Request a refund for a succeeded payment (recorded, not auto-executed — see docs/security.md)
      parameters:
        - $ref: '#/components/parameters/PaymentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount]
              properties:
                amount: { type: integer }
                reason: { type: string }
      responses:
        '201':
          description: Refund request recorded
        '409':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    PaymentId:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
  responses:
    Error:
      description: Standard error envelope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success: { type: boolean, example: false }
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }